+++ R/generateSeqDates.R
@@ -1,11 +1,15 @@
 #' POSIXct date sequence generator
 #'
-#' Simple helper function that generates a simple sequence of POSIXct dates between a specified starting year and an ending year. The frequency can be specified. The date sequence starts at 00:00:00.
+#' Simple helper function that generates a simple sequence of POSIXct dates
+#' between a specified starting year and an ending year. The frequency can be
+#' specified. The date sequence starts at 00:00:00.
 #'
 #' @param startY Starting year for sequence of dates
 #' @param endY Ending year for sequence of dates
 #' @param freq "hour", "day", "week", "month", "quarter" or "year"
-#' @return Tibble with date sequence with date frequency as specified. The date sequence starts at 00:00:00
+#' @param tz time zone. Default UTC.
+#' @return Tibble with date sequence with date frequency as specified. The date
+#'   sequence starts at 00:00:00
 #' @family Helper functions
 #' @export
 generateSeqDates <- function(startY,endY,freq,tz="UTC"){