+++ R/plotNormDevHYY.R
@@ -11,19 +11,24 @@
 #' @param dataType String, either 'Q' for mean discharge in m3/s, 'T' for temperature in degrees Celsius or 'P' for precipitation in mm/yr.
 #' @param stationID String of station name, e.g. 'Khorog-Gunt 17050'
 #' @return ggplot object
+#' @importFrom rlang .data
 #' @family Pre-processing
 #' @export
 plotNormDevHYY <- function(data2Plot,dataType,stationID){
-  # rename columns just to ensure proper functioning of the function after changes in the naming convention of convert2HYY()
+
+  . <- NULL
+
+  # rename columns just to ensure proper functioning of the function after
+  # changes in the naming convention of convert2HYY()
   data2Plot <- data2Plot %>% dplyr::rename(data = 2, data_cs = 3, data_ws = 4)
   # augment data where data is hydrological year time series object
   data_norms <- data2Plot %>%
-    dplyr::mutate(norm_data = base::mean(data,na.rm = TRUE)) %>%
-    dplyr::mutate(norm_data_cs = base::mean(data_cs, na.rm = TRUE)) %>%
-    dplyr::mutate(norm_data_ws = base::mean(data_ws, na.rm = TRUE)) %>%
-    dplyr::mutate(delta_data = data - norm_data) %>%
-    dplyr::mutate(delta_data_cs = data_cs - norm_data_cs) %>%
-    dplyr::mutate(delta_data_ws = data_ws - norm_data_ws)
+    dplyr::mutate(norm_data = base::mean(.data$data,na.rm = TRUE)) %>%
+    dplyr::mutate(norm_data_cs = base::mean(.data$data_cs, na.rm = TRUE)) %>%
+    dplyr::mutate(norm_data_ws = base::mean(.data$data_ws, na.rm = TRUE)) %>%
+    dplyr::mutate(delta_data = .data$data - .data$norm_data) %>%
+    dplyr::mutate(delta_data_cs = .data$data_cs - .data$norm_data_cs) %>%
+    dplyr::mutate(delta_data_ws = .data$data_ws - .data$norm_data_ws)
   # generate titles
   if (dataType == 'Q') {
     ylabText = 'Q [m3/s]'
@@ -52,23 +57,29 @@ plotNormDevHYY <- function(data2Plot,dataType,stationID){
   }
   # plot deviation from norm for hydrological year data using the geom_step function
   completeHYY <- data_norms %>%
-    ggplot2::ggplot(ggplot2::aes(x = hyYear, y = delta_data)) +
-    ggplot2::geom_bar(ggplot2::aes(fill = delta_data < 0), stat = "identity") +
-    ggplot2::scale_fill_manual(guide = FALSE, breaks = c(TRUE, FALSE), values = c("red", "blue")) +
+    ggplot2::ggplot(ggplot2::aes(x = .data$hyYear, y = .data$delta_data)) +
+    ggplot2::geom_bar(ggplot2::aes(fill = .data$delta_data < 0), stat = "identity") +
+    ggplot2::scale_fill_manual(guide = "none", breaks = c(TRUE, FALSE),
+                               values = c("red", "blue")) +
     ggplot2::xlab('Hydrological year') + ggplot2::ylab(ylabText) +
-    ggplot2::ggtitle(titleText_HYY)
+    ggplot2::ggtitle(titleText_HYY) +
+    ggplot2::theme_bw()
   csHYY <- data_norms %>%
-    ggplot2::ggplot(aes(x = hyYear, y = delta_data_cs)) +
-    ggplot2::geom_bar(aes(fill = delta_data_cs < 0), stat = "identity") +
-    ggplot2::scale_fill_manual(guide = FALSE, breaks = c(TRUE, FALSE), values = c("red", "blue")) +
+    ggplot2::ggplot(ggplot2::aes(x = .data$hyYear, y = .data$delta_data_cs)) +
+    ggplot2::geom_bar(ggplot2::aes(fill = .data$delta_data_cs < 0), stat = "identity") +
+    ggplot2::scale_fill_manual(guide = "none", breaks = c(TRUE, FALSE),
+                               values = c("red", "blue")) +
     ggplot2::xlab('Hydrological year') + ggplot2::ylab(ylabText) +
-    ggplot2::ggtitle(titleText_cs)
+    ggplot2::ggtitle(titleText_cs) +
+    ggplot2::theme_bw()
   wsHYY <- data_norms %>%
-    ggplot2::ggplot(ggplot2::aes(x = hyYear, y = delta_data_ws)) +
-    ggplot2::geom_bar(ggplot2::aes(fill = delta_data_ws < 0), stat = "identity") +
-    ggplot2::scale_fill_manual(guide = FALSE, breaks = c(TRUE, FALSE), values = c("red", "blue")) +
+    ggplot2::ggplot(ggplot2::aes(x = .data$hyYear, y = .data$delta_data_ws)) +
+    ggplot2::geom_bar(ggplot2::aes(fill = .data$delta_data_ws < 0), stat = "identity") +
+    ggplot2::scale_fill_manual(guide = "none", breaks = c(TRUE, FALSE),
+                               values = c("red", "blue")) +
     ggplot2::xlab('Hydrological year') + ggplot2::ylab(ylabText) +
-    ggplot2::ggtitle(titleText_ws)
+    ggplot2::ggtitle(titleText_ws) +
+    ggplot2::theme_bw()
 
   ggpubr::ggarrange(completeHYY, csHYY, wsHYY, ncol = 1) %>% return()
 }